일정 금액 이상 올라가면 Gold 초기화되는 버그 디버깅 완료, 상호작용에 따라 Inputmode 수정

+++ VICTIMS/Source/VICTIMS/Public/AVICTIMSPlayerController.h
@@ -61,7 +61,7 @@ public:
 	virtual void UI_SellItem_Implementation(const uint8& InventorySlot) override;
 	/* Ends Interface */
 
-	uint8 UIGetPlayerGold();
+	int32 UIGetPlayerGold();
 
 	UPROPERTY(VisibleAnywhere, BlueprintReadWrite, meta=(Category="Inventory", OverrideNativeName="InventoryComponent"))
 	UInventoryManagerComponent* InventoryManagerComponent;
